Add link to latest User Guide on front page
The user guide is really the most important piece of information to get new folks on board to test out CWL. Even though probably having stumbled upon the user guide before, I wasn't aware that it existed, and did not find it when I started evaluating CWL for use within NBIS here in Sweden. I thus think it is really important to have a link to it clearly visible in a prominent place on the front page.